Break apart the pipboy app into pieces.
