+++ index.html
@@ -51,6 +51,8 @@
     <div id="arrow">
       <div class="inner"></div>
     </div>
-    <script src="src/index.js"></script>
+    <script>
+      require('./src/index')();
+    </script>
   </body>
 </html>
